:bug:[#283] 여행 정렬 기준을 글로벌 클로저 travelModalSortRule로 분리
