+++ UMM/View/Modal/AddTravelRequestModal.swift
@@ -25,7 +25,7 @@ struct AddTravelRequestModal: View {
     }
     
     private var descriptionView: some View {
-        Text("현재 진행 중인 여행이 없어요.\n여행을 생성할까요?")
+        Text("현재 생성된 여행이 없어요.\n여행을 생성할까요?")
             .foregroundStyle(.black)
             .font(.display1)
             .multilineTextAlignment(.center)
