+++ UMM/ViewModel/MainViewModel.swift
@@ -11,6 +11,10 @@ import SwiftUI
 class MainViewModel: ObservableObject {
     
     static let shared = MainViewModel()
+    var timer = Timer.scheduledTimer(withTimeInterval: 3.0, repeats: true) { timer in
+        print("MainViewModel | selectedTravel: \(shared.selectedTravel?.name ?? "nilName")")
+        print("MainViewModel | selectedTravelInExpense: \(shared.selectedTravelInExpense?.name ?? "nilName")")
+    }
     
     // didSet으로 selection == 2일 때, default
     @Published var selection: Int = 1
@@ -18,11 +22,7 @@ class MainViewModel: ObservableObject {
         didSet {
             if selectedTravel?.name != tempTravelName {
                 if selectedTravel != selectedTravelInExpense {
-                    if let selectedTravel = selectedTravel, selectedTravel.name == tempTravelName {
-                        self.selectedTravelInExpense = findCurrentTravel()
-                    } else {
-                        updateSelectedTravelInExpense()
-                    }
+                    updateSelectedTravelInExpense()
                 }
             }
         }
